### 2015-12-24 #23 @tonyfischetti
Adding `assertr` to ROpenSci
* 1. What does this package do? (explain in 50 words or less)
The assertr package supplies a suite of functions designed to verify assumptions about data early in an analysis pipeline to protect against common data errors and instances of bad data.

* 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).
```
Package: assertr
Title: Assertive Programming for R Analysis Pipelines
Version: 1.0.0
Authors@R: person("Tony", "Fischetti", email="tony.fischetti@gmail.com",
  role = c("aut", "cre"))
Maintainer: Tony Fischetti <tony.fischetti@gmail.com>
Description: Provides functionality to assert conditions
    that have to be met so that errors in data used in
    analysis pipelines can fail quickly. Similar to
    'stopifnot()' but more powerful, friendly, and easier
    for use in pipelines.
    MASS,
    lazyeval
    magrittr
```

* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/tonyfischetti/assertr

* 4. What data source(s) does it work with (if applicable)?
Any. Mostly in the form of `data.frames`

* 5. Who is the target audience?
Anyone who has ever struggled with bad data

* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
The `ensurer` package attempts to solve the very same problem. To a certain extent, the `assertive` package also offer some similar capabilities. The difference between `assertr` and these other packages is the grammar of usage and the way that assertions of different types can be easily combined to express arbitrarily complex assertions in a very readable way.

* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
* [x] This package does not violate the Terms of Service of any service it interacts with.
* [x] The repository has continuous integration with Travis and/or another service
* [x] The package contains a vignette
* [x] The package contains a reasonably complete readme with devtools install instructions
* [x] The package contains unit tests
* [x] The package only exports functions to the NAMESPACE that are intended for end users

* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
Yes. All the user-facing functions are in snake case, but the internal functions sometimes use dots (`.`) as separators. I'm open to changing that. Also, the package doesn't have a code of conduct yet but I think it's a good idea to include.

* [ ] Are there any package dependencies not on CRAN?
No

It already is

* [x] Does the package have a CRAN accepted license?
You bet

* [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.
No warnings

* 9. Please add explanations below for any exceptions to the above:

* 10. If this is a resubmission following rejection, please explain the change in cirucmstances.
