### 2015-12-24 #23 @richfitz
I have a use case for this and have already looked through the code, so am happy to review if that is useful.
