### 2016-01-22 #23 @richfitz
The descent towards manuscript central begins... :grin:
