### 2016-01-22 #23 @jennybc
Maybe we should send hand-written notes?
