### 2016-01-22 #23 @sckott
> Definitely not! I think it's great. I actually thought it was you, which you could never say for MS central.

`heythere` != MS central

> Maybe we should send hand-written notes?

Yes!
