### 2016-01-25 #23 @richfitz
# General comments
​
The `assertr` package provides a generalised framework for defensive programming around data.frames.  This sits somewhere between `stopifnot` and `testthat` in terms of flexibility and complexity and as such forms a useful building block for data analysis workflow, which I believe is under-tooled at the moment.  I really like the idea of having packages that are primarily focussed on data workflows rather than restricting people to ideas that were developed for software engineering (such as formal unit tests).
​
The package is very tight -- it exports the minimum set of functionality and conforms to the "do one thing and do it well" school of thought.  The functions are well documented, the vignette is readable and less dry than most.  I appreciate the split into NSE and SE versions of all core functions.
​
Accordingly, most of my comments focus on design decisions and therefore may all be out of line because the author will have thought about this more than I have.
​
**The main entrypoints are difficult to differentiate**
​
My biggest concern is that I found the three main entry points very difficult to keep straight.  And when I put the package down over Christmas I had to re-remember them again.
​
* `assert(data, predicate, ...)`
* `verify(data, expr, ...)`
* `insist(data, predicate_generator, ...)`
​
The difference is primarily in the properties of the second argument and I wonder if there's a way of specifying that some other way than three functions that have such similar names?  The current approach is extremely elegant but at the cost of being a bit too opaque to the user -- especially because the three function names are essentially synonyms of each other there is nothing to jog your memory.  I presume the problem is it is difficult to detect the difference between the three argument types before they are evaluated (and the correct evaluation depends on the type).
​
**The custom handler routine is inflexible**
​
The custom handler is a really great addition, but could be improved.  `testthat` has a similar handler approach that allows storage of a bunch of repeated assertions (pass or fail).  My use-case for this package is to replicate something like this, so I'd want to pass the same handler in to all the functions in a pipeline:
​
```
mtcars %>%
  verify(nrow(mtcars) > 10, error_fun=my_error_fun) %>%
  verify(mpg > 0, error_fun=my_error_fun) %>%
  insist(within_n_sds(4), mpg, error_fun=my_error_fun) %>%
  assert(in_set(0,1), am, vs, error_fun=my_error_fun) %>%
  group_by(cyl) %>%
  summarise(avg.mpg=mean(mpg))
```
​
What would be heaps nicer is if I could register a handler; change the `assert` function to something like:
​
```
assertr <- function(data, predicate, ..., error_fun=getOption("assertr.handler", assertr_stop)) {
}
```
​
(or eqivalently use the package-environment trick like `testthat` does).  This would allow:
​
```
options(assertr.handler=my_error_fun)
mtcars %>%
  verify(nrow(mtcars) > 10) %>%
  verify(mpg > 0) %>%
  insist(within_n_sds(4), mpg) %>%
  assert(in_set(0,1), am, vs) %>%
  group_by(cyl) %>%
  summarise(avg.mpg=mean(mpg))
```
​
(As a related comment, the usage definitions reference the unexported function `assertr_stop` which some may find confusing.  Additionally, is there a reason why `verify` uses `error_fun=stop` not `assertr_stop`?)
​
# Minor comments
​
* **The `dplyr` dependency**, which is used soley for `dplyr::select_` seems a potentially heavy dependency for one function; if it is straightforward to swap out for independent imlementation that would decrease the package footprint (I can imagine using this in contexts where I do not have dplyr installed such as container-based workflows).  But I can totally see the advantage of sticking with something that is known to work.
​
* **Think about the non-pipe people**.  All the examples and the vignette make heavy use of the `%>%`operator (which is fine), but as someone who uses this little or who imagines using this mostly in packages where I'd be avoiding so much weird evaluation, I would appreciate a few more pipe-less examples.  This is potentially confusing in places like:
​
```
our.data %>% assert(within_bounds(0,Inf), mpg) # and so on
```
​
when `?assert` says that usage is
​
```
assert(data, predicate, ..., error_fun = assertr_stop)
```
​
This confused me because in usage it looks like `predicate, data`, but of course the data argument comes from the pipe and the `mpg` is the column name is passed through to the `...` argument.  While the use with pipes is very elegant, I think the package has use outside of that scope too.  The examples within the package are actually really good like this.
​
* **Classed exceptions would make the error handling more flexible**.  Related to the second major point above; it would be nice to distinguish between errors that were because the input to assertr was incorrect, and errors that are raised because the data failed the assertions.  R's classed errors provide a nice framework for this.  Then `tryCatch` and `withCallingHandlers` can dispatch appropriately based on the sort of error.
​
* **A reporting framework would be fantastic** If you don't write this, I will -- but given you have written this package I figure you should get right of refusal.  Related to the point above, I would like to use the underlying bits you have here in a package for automated testing of upstream data sources that tend to be misbehaved.  I can imagine a testthat-like reporting framework where a bunch of tests are run and the failures reported.
Add Comment
