### 2016-02-20 #23 @tonyfischetti
Thanks for the kind words and really great feedback @richfitz.

> The main entrypoints are difficult to differentiate
As pointed out, this is somewhat a consequence of my making the API as elegant as possible but at the expense of some opaque-ness. The problem is (again, as you pointed out) there's no elegant way that I can see to programmatically detect the argument types. So we have verbs that are *literally* synonyms (I chose their names using a thesaurus). Unfortunately, I can't think of a better naming mechanism without really long function names like `take_a_predicate_generator_and_apply_to_each_column()`. I'm open to suggestions, but it may not be the end of the world since there are only three main verbs and the docs are good.

> The custom handler routine is inflexible
I like the idea of using a `testthat` style `options` mechanism. I'll implement this

> Additionally, is there a reason why verify uses error_fun=stop not assertr_stop?)
Nope, that's an error. Good catch!

> The dplyr dependency
That was a difficult choice. It's a *hell* of a heavy dependency, but I was wary of implementing `dplyr::select` myself. Especially since I would have to reference `dplyr`'s implementation so heavily that it would likely constitute code theft. I'd love to drop the dependency though, if anyone thinks they can implement it without copying code.

> A reporting framework would be fantastic
I *need* this feature, and I have a few great ideas on how to implement it. I'd like to talk to you more (@richfitz) about your particular use case in case my solution only suits my use case. I think this feature can potentially be the most powerful and useful capability of `assertr`
