### 2016-03-12 #23 @tonyfischetti
So I'm going to run into a little more free time in the near future and I'd like to get back from my learning hiatus back into improving `assertr`--particularly because people are telling me its *really* useful for them. Because of the learning hiatus, I have some fresh new ideas for improvement, but I'd like to run them past some of you for further input...

> The main entrypoints are difficult to differentiate

As mentioned before, there is `assert`, `insist`, `insist_rows`, and `assert_rows` for representing a wide range of tasks. However, if I wanted to add the ability to, say, declare that the whole data set should have no more than 15% missing values (and I do want to add that), the semantics of that would require *another* specialized function... and I'm running out of synonyms for "assert"!

Principled though that solution was, I'm not sure its the correct thing to do going forward; it's always been one of R's strengths from a user's point of view to use a familiar generic function (`mean`, `plot`, etc...) with all sorts of input and have the object system dynamically dispatch the correct functionality.

So how about this... creating an ~~S3~~S4 generic function (`proclaim` perhaps?) that will handle all of the different semantics *for* the user. Concretely, the function returned from `within_n_mads` can be labeled with class `assertr_dynamic` (because the predicate is dynamically generated). Then, `proclaim` would dispatch on the *second* argument (the first arg is the data frame) and call what is currently referred to as `insist`. In the same way, `maha_dist` would be classed something like `assertr_dynamic_rows` and a user calling `proclaim(df, maha_dist, within_n_mads(10), ...)` would transparently dispatch `insist_rows(maha_dist, within_n_mads(10), ...)`

This would improve `assertr`'s extensibility greatly; for example, adding semantics to check the supplied data.frame *as a whole* would only require writing another ~~S3~~S4 method of the `proclaim` generic and making sure that the predicate function was correctly classed.

This solution is perhaps a little unconventional, but it completely obviates the requirement that a useR remember all the verbs.

> The custom handler routine is inflexible

The most common suggestion for `assertr` is to be able to warn (not error) on violation. Additionally, even if it does error on violation, there should be semantics in place for the entire chain of assertions to run so that the final error message will contain the complete report of the data errors.

The reason I needed to take my time with this one is that I need the warnings to be concatenated through a `assertr` chain in a principled manner. To do this without using dynamic variables (eww), it requires that--along with returned the data frame given--the `assertr` verbs need to return the warnings so that they can be concatenated with the warnings further in the assertion pipeline. Up until recently, I thought that I would have to implement something that would be tantamount to a Haskell monad in order to do this. Another possibility is to use S4 (hence the ~~S3~~ strikeouts in the paragraphs above) in order to get `proclaim` (or whatever it is) to dispatch on *both* the second argument, and the *first* argument. If the first argument is a `data.frame` the current semantics stand... if the first argument is something else (some class that holds a `data.frame` and a running list of errors) then the wanted behavior can be dispatched.

The big complication is what happens at the end of the chain. There needs to be something that tells `assertr` that the chain is ending so that it can take the `data.frame` out of the composite data.frame/error_log object and finally display the error or warning. Any ideas?

I'd appreciate any feedback on these ideas for two reasons (a) it's now (or will hopefully be soon) ROpenSci's project not just mine, and (b) I'd like to get the input of some talented developers :)
