### 2016-03-12 #23 @tonyfischetti
Also, none of the proposed additions have to break backwards compatibility :)
