### 2016-03-12 #23 @jennybc
Since I'm the one who is yet to review ... @tonyfischetti do you have recommendations of a good dataset to run through `assertr`? I.e. one that you think should show it off but ... there's enough uncertainty that it would be interesting to see how things go? Also to see how a new user manages with it. I have one idea that I'll fall back on if nothing immediately comes to mind.
