### 2016-03-12 #23 @tonyfischetti
@jennybc Nothing *immediately* comes to mind but I'm sure I can dig up one of the examples that inspired me to get into this package in the first place :)
