### 2016-03-13 #23 @aammd
@jennybc wellll i would not say _ugly_, but rather "pre-assertr". it represents "how we did this before assertr" and highlights improvements in the UI of this package (improvements I tried to show in [my lesson about the `assertr`](http://biol548o.github.io/assert.html))
