### 2016-05-25 #23 @tonyfischetti
@smbache I didn't know `ensurer` was being considered
