### 2016-06-08 #23 @sckott
@tonyfischetti approved!

-   Add the footer to your README:

```
[![ropensci\_footer](http://ropensci.org/public_images/github_footer.png)](http://ropensci.org)
```

- Update installation of dev versions to `ropenscilabs/assertr` and any urls for the github repo to `ropenscilabs` instead of `tonyfischetti`
- Update any links to the package from `tonyfischetti/assertr` to `ropenscilabs/assertr` (though even if they aren't github will redirect to the new location :) )
- Go to the Repo Settings --> Transfer Ownership and transfer to `ropenscilabs` - Note that all our newer pkgs go to `ropenscilabs` first, then when more mature we'll move to `ropensci`
