### 2016-06-14 f4daf20
attempting to fix travis ci integration problems
