### 2016-06-14 563b3ed
trying again to fix travis problems
