### 2016-06-14 7945d6c
fixed issue #25 and made it clear that non-factor character columns...
are not automatically converted to factors by maha_dist
