### 2016-06-14 b0d02d5
nary a NOTE in sight
