+++ R/predicates.R
@@ -226,7 +226,7 @@ within_n_sds <- function(n, ...){
     if(!is.vector(a.vector) || !is.numeric(a.vector))
       stop("argument must be a numeric vector")
     mu <- mean(a.vector, na.rm=TRUE)
-    stdev <- sd(a.vector, na.rm=TRUE)
+    stdev <- stats::sd(a.vector, na.rm=TRUE)
     if(is.na(mu)) stop("mean of vector is NA")
     if(is.na(stdev)) stop("standard deviations of vector is NA")
     within_bounds((mu-(n*stdev)), (mu+(n*stdev)), ...)
@@ -294,8 +294,8 @@ within_n_mads <- function(n, ...){
   function(a.vector){
     if(!is.vector(a.vector) || !is.numeric(a.vector))
       stop("argument must be a numeric vector")
-    dmad <- mad(a.vector, na.rm=TRUE)
-    dmed <- median(a.vector, na.rm=TRUE)
+    dmad <- stats::mad(a.vector, na.rm=TRUE)
+    dmed <- stats::median(a.vector, na.rm=TRUE)
     if(is.na(dmad)) stop("MAD of vector is NA")
     if(is.na(dmed)) stop("median of vector is NA")
     within_bounds((dmed-(n*dmad)), (dmed+(n*dmad)), ...)
