+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: assertr
 Type: Package
 Title: Assertive Programming for R Analysis Pipelines
-Version: 1.0.0
+Version: 1.0.1
 Authors@R: person("Tony", "Fischetti", email="tony.fischetti@gmail.com",
   role = c("aut", "cre"))
 Maintainer: Tony Fischetti <tony.fischetti@gmail.com>
@@ -17,7 +17,8 @@ LazyData: TRUE
 Imports:
     dplyr,
     MASS,
-    lazyeval
+    lazyeval,
+    stats
 Suggests:
     knitr,
     testthat,
