### 2016-06-14 dd3e839
preparing for transfer of ownership
